Rename corkboard folder to docs
