+++ .gitignore
@@ -1,6 +1,5 @@
 node_modules/
 coverage/
 dist/
-docs/
 npm-debug.log
 stats.html
