+++ docs/src/components/Checkerboard.js
