+++ src/GroupAvatar/GroupAvatar.doc.js
@@ -8,7 +8,7 @@ import {
   PropTable,
   Example,
   Combination,
-} from '../../.corkboard/src/cards';
+} from '../../docs/src/cards';
 
 ns(
   'Group Avatar',
