+++ src/Heading/Heading.doc.js
@@ -1,6 +1,6 @@
 // @flow
 import * as React from 'react';
-import { ns, card, md, PropTable, Example } from '../../.corkboard/src/cards';
+import { ns, card, md, PropTable, Example } from '../../docs/src/cards';
 
 ns(
   'Heading',
