+++ src/Image/Image.doc.js
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react';
 import Image from './Image';
-import { ns, card, md, PropTable, Example } from '../../.corkboard/src/cards';
+import { ns, card, md, PropTable, Example } from '../../docs/src/cards';
 
 ns(
   'Image',
